[REFACTOR]: 사업자번호 검증 코드 패키지명 변경
